+++ src/main/java/com/project/baedalsodae/event/dto/OrderCreatedEvent.java
@@ -1,9 +1,10 @@
 package com.project.baedalsodae.event.dto;
 
 import com.project.baedalsodae.order.entity.Order;
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record OrderCreatedEvent(UUID orderId, UUID userId, int finalAmount) {
+public record OrderCreatedEvent(UUID orderId, UUID userId, BigDecimal finalAmount) {
     public static OrderCreatedEvent from(Order order) {
         return new OrderCreatedEvent(order.getId(), order.getUserId(), order.getFinalAmount());
     }
